[backend] Rename OS to platform
