+++ openex-api/src/main/resources/application.properties
@@ -179,3 +179,4 @@ collector.users.enable=false
 collector.caldera.enable=false
 collector.caldera.url=http://localhost:8888
 collector.caldera.api-key=ChangeMe
+collector.caldera.interval=60
